meson: fix regression that broke extracting version (#1253)
* meson: fix regression that broke extracting version

In commit 6c0da1604709e53a4ce4de02d2b0abcacbb24140 the code that
heuristically parsed the version broke due to the version being moved
around into a more easily accessible define.

While we are at it, pass the exact path of httplib.h to un-break usage
as a meson subproject. This was broken in commit
7f06e22b018f9323aafc631c73b5bcd12c284b7a which checked the return code
of trying to get the version; it was always broken, but formerly failed
in silence and resulted in no version number.

* meson: use the compiler builtins to extract the version from the header

As a convenient string define, it is now possible to ask the
preprocessor what the version of cpp-httplib is. This can be used from
meson too, in order to avoid encoding C++ file structure into python
regexes.